Add --host option for serving and use Flask default if not specified
As mentioned in Flask docs, "0.0.0.0" should be used for externally
visible servers. It is not a good default for security reasons.

Change the default to Flask's own default (127.0.0.1), and provide
a --host option to change it.

Unfortunately, there's no easy way to test the difference between
binding to 127.0.0.1 and 0.0.0.0. That would requite knowing an
external IP of the current machine, and possibly opening up firewalls.